[#51] Fix: DiscountPolicy의 할인금액 계산로직 반환타입 int로 변경

+++ BE/src/main/java/codesquad/sidedish/domain/discount/DiscountPolicy.java
@@ -29,7 +29,7 @@ public class DiscountPolicy {
         this.discountPolicyId = discountPolicyId;
     }
 
-    public long calculateDiscountedPrice(long price) {
-        return (long)(price * (1 - discountRate));
+    public int calculateDiscountedPrice(int price) {
+        return (int)(price * (1 - discountRate));
     }
 }
